refactor: 멀티 액티비티 구조를 위해 플랫폼 진입점을 각 feature 모듈로 이동
각 feature 모듈이 자체 Activity/ViewController를 소유하도록 변경하여
모듈별 독립적인 플랫폼 진입점을 갖는 멀티 액티비티 구조 기반 마련
