+++ app-android/src/main/kotlin/com/linkit/company/android/LinkitApplication.kt
@@ -3,7 +3,7 @@ package com.linkit.company.android
 import android.app.Application
 import android.content.Context
 import com.linkit.company.AndroidAppGraph
-import com.linkit.company.AppGraph
+import com.linkit.company.core.common.AppGraph
 import dev.zacsweers.metro.createGraphFactory
 
 class LinkitApplication : Application() {
