--- app-shared/src/commonMain/kotlin/com/linkit/company/AppGraph.kt
@@ -1,3 +1,0 @@
-package com.linkit.company
-
-interface AppGraph
