+++ core/common/src/commonMain/kotlin/com/linkit/company/core/common/AppGraph.kt
@@ -1,0 +1,3 @@
+package com.linkit.company.core.common
+
+interface AppGraph
