+++ feature/home/src/iosMain/kotlin/com/linkit/company/feature/home/HomeViewController.kt
@@ -1,13 +1,12 @@
-package com.linkit.company
+package com.linkit.company.feature.home
 
-import androidx.compose.runtime.Composable
+import androidx.compose.ui.window.ComposeUIViewController
+import com.linkit.company.core.common.AppGraph
 import com.linkit.company.core.designsystem.theme.LinkItTheme
 import com.linkit.company.core.navigation.LinkItSavedStateConfiguration
-import com.linkit.company.feature.home.HomeScreen
 
-@Composable
-context(appGraph: AppGraph)
-fun LinkItApp() {
+@Suppress("UNUSED")
+fun HomeViewController(appGraph: AppGraph) = ComposeUIViewController {
     LinkItTheme {
         HomeScreen(savedStateConfiguration = LinkItSavedStateConfiguration)
     }
