+++ feature/intro/src/iosMain/kotlin/com/linkit/company/feature/intro/IntroViewController.kt
@@ -1,8 +1,7 @@
-package com.linkit.company
+package com.linkit.company.feature.intro
 
 import androidx.compose.ui.window.ComposeUIViewController
 import com.linkit.company.core.designsystem.theme.LinkItTheme
-import com.linkit.company.feature.intro.IntroScreen
 
 fun IntroViewController(onComplete: () -> Unit) = ComposeUIViewController {
     LinkItTheme {
